refactor: extract DebugStackTrace from StackTraceFactory
