+++ src/Sentry/Extensibility/ISentryStackTraceFactory.cs
@@ -13,10 +13,4 @@ public interface ISentryStackTraceFactory
     /// <param name="exception">The exception to create the stacktrace from.</param>
     /// <returns>A Sentry stack trace.</returns>
     SentryStackTrace? Create(Exception? exception = null);
-
-    /// <summary>
-    /// Returns a list of <see cref="DebugImage" />s referenced from the previously processed <see cref="Exception" />s.
-    /// </summary>
-    /// <returns>A list of referenced debug images.</returns>
-    List<DebugImage>? DebugImages() { return null; }
 }
